# Further reduce ensureAccessible() overhead