@njhill 2019-02-27
Thanks @normanmaurer @franz1981 @ejona86, have added a commit addressing the comments.