@normanmaurer 2019-02-28
@netty-bot test this please